+++ src/wfg/native_ui/ui/component/LayoutOffset.java
@@ -1,6 +1,6 @@
 package wfg.native_ui.ui.component;
 
-public final class LayoutOffset extends BaseComponent {
+public final class LayoutOffset {
     public int x, y, w, h;
 
     public final void setOffset(int x, int y, int w, int h) {