+++ src/wfg/native_ui/ui/widget/RadioPanel.java
@@ -124,7 +124,7 @@ public class RadioPanel extends CustomPanel implements UIBuildableAPI {
                 buttons.add(button);
 
                 final float x = pad + i * (buttonWidth + pad);
-                add(button).inTL(x, pad);
+                add(button).inTL(x, 0f);
 
                 if (i == 0) button.cutStyle = CutStyle.TL_BL;
                 if (i == count - 1) button.cutStyle = CutStyle.TR_BR;